requirements-dev.txt
black==19.3b0
pre-commit==1.18.3
pytest==5.2.1
# enable local airflow dag development
apache-airflow
psycopg2-binary
